+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-log4j-2.x-activation/src/main/java/org/apache/skywalking/apm/toolkit/activation/log/log4j/v2/x/async/AsyncLoggerConfigDisruptorInterceptor.java
@@ -18,41 +18,31 @@
 
 package org.apache.skywalking.apm.toolkit.activation.log.log4j.v2.x.async;
 
-import org.apache.skywalking.apm.agent.core.context.ContextManager;
+import java.lang.reflect.Method;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
-import org.apache.skywalking.apm.toolkit.logging.common.log.SkyWalkingContext;
 
-import java.lang.reflect.Method;
-
-/**
- * <p>Pass the global trace context into the _sw field of Lo4jLogEvent instance after enhancing</p>
- */
-
-public class AsyncLoggerConfigMethodInterceptor implements InstanceMethodsAroundInterceptor {
+public class AsyncLoggerConfigDisruptorInterceptor implements InstanceMethodsAroundInterceptor {
 
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
                              MethodInterceptResult result) throws Throwable {
-        // LogEvent has not been enhanced
-        if (allArguments[0] instanceof EnhancedInstance && ((EnhancedInstance) allArguments[0]).getSkyWalkingDynamicField() == null) {
-            SkyWalkingContext skyWalkingContext = new SkyWalkingContext(
-                ContextManager.getGlobalTraceId(), ContextManager.getSegmentId(), ContextManager.getSpanId());
-
-            EnhancedInstance instances = (EnhancedInstance) allArguments[0];
-            instances.setSkyWalkingDynamicField(skyWalkingContext);
-        }
     }
 
     @Override
     public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
-        Object ret) throws Throwable {
+                              Object ret) throws Throwable {
+        if (ret instanceof EnhancedInstance && allArguments[0] instanceof EnhancedInstance) {
+            EnhancedInstance oldInstance = (EnhancedInstance) allArguments[0];
+            EnhancedInstance newInstance = (EnhancedInstance) ret;
+            newInstance.setSkyWalkingDynamicField(oldInstance.getSkyWalkingDynamicField());
+        }
         return ret;
     }
 
     @Override
     public void handleMethodException(EnhancedInstance objInst, Method method, Object[] allArguments,
-        Class<?>[] argumentsTypes, Throwable t) {
+                                      Class<?>[] argumentsTypes, Throwable t) {
     }
 }