+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-log4j-2.x-activation/src/main/java/org/apache/skywalking/apm/toolkit/activation/log/log4j/v2/x/async/LogEventMethodInterceptor.java
@@ -27,27 +27,27 @@ import org.apache.skywalking.apm.toolkit.logging.common.log.SkyWalkingContext;
 import java.lang.reflect.Method;
 
 /**
- * <p>Pass the global trace context into the _sw field of RingBufferLogEvent instance after enhancing</p>
+ * <p>Pass the global trace context into the _sw field of LogEvent instance after enhancing</p>
  */
 
-public class RingBufferLogEventMethodInterceptor implements InstanceMethodsAroundInterceptor {
+public class LogEventMethodInterceptor implements InstanceMethodsAroundInterceptor {
 
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
-        MethodInterceptResult result) throws Throwable {
-        SkyWalkingContext skyWalkingContext = new SkyWalkingContext(ContextManager.getGlobalTraceId(),
-                ContextManager.getSegmentId(), ContextManager.getSpanId());
+                             MethodInterceptResult result) throws Throwable {
+        SkyWalkingContext skyWalkingContext = new SkyWalkingContext(
+            ContextManager.getGlobalTraceId(), ContextManager.getSegmentId(), ContextManager.getSpanId());
         objInst.setSkyWalkingDynamicField(skyWalkingContext);
     }
 
     @Override
     public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
-        Object ret) throws Throwable {
+                              Object ret) throws Throwable {
         return ret;
     }
 
     @Override
     public void handleMethodException(EnhancedInstance objInst, Method method, Object[] allArguments,
-        Class<?>[] argumentsTypes, Throwable t) {
+                                      Class<?>[] argumentsTypes, Throwable t) {
     }
 }